Refactor config handling and fix mistakes in Utils.pm
Refactored the configuration handling in extract_libraryiq.pl by updating the location of the config file and altering how configurations are read and returned as a reference. Adjusted the read_config function in Utils.pm to remove unused parameters and return a hash reference. Corrected a typo in the get_db_config structure access.

- Updated config file path in extract_libraryiq.pl
- Modified read_config to return a hash reference
- Removed logging from read_config in Utils.pm as it was unused
- Corrected hash access typo in get_db_config

Release-Note: Refactor config handling in extract_libraryiq and fix hash access error

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>