Remove redundant condition from circulation query
Simplified the SQL query in the get_circ_ids_sql method by removing the unnecessary condition that checks for non-deleted records in the circulation table. The existence of this condition was redundant as the join with the asset tables already ensures integrity.

Release-Note: Simplifies SQL query by removing redundant condition from circulation checks

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>