Ensure UTF-8 encoding when writing data to file
Corrects the encoding format for data written to files to UTF-8, which prevents issues with special or non-ASCII characters. This change ensures that data integrity is maintained when reading the output files on systems or applications that expect UTF-8 encoded content.

Release-Note: Ensure file writing in UTF-8 encoding to maintain character integrity

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>